+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - FiFi SDR: prevent arecord from shutting down after 2GB of data has been sent
 - Added support for bias tee control on rtl_sdr devices
 - All connector driven SDRs now support `"rf_gain": "auto"` to enable AGC
+- `rtl_sdr` type now also supports the `direct_sampling` option
 
 **0.18.0**
 - Support for SoapyRemote